+++ src/process_state.h
@@ -11,16 +11,16 @@ private:
 public:
   // last timestamp process was executed on cpu.
   // initialized to process arrival time
-  int m_LastCpuTime;
+  double m_LastCpuTime;
   // total time process has spent executing
-  int m_TotalCpuTime = 0;
+  double m_TotalCpuTime = 0;
   // time between process arrival and completion
-  int m_TurnaroundTime = 0;
+  double m_TurnaroundTime = 0;
   // total time process has spent waiting
-  int m_WaitingTime = 0;
+  double m_WaitingTime = 0;
 
   ProcessState();
-  ProcessState(int pid, int arrival_time);
+  ProcessState(int pid, double arrival_time);
   int getPid() const { return m_Pid; }
 
   friend std::ostream &operator<<(std::ostream &os, const ProcessState &obj);
